requirements.txt
email_validator==1.3.0
Flask==2.2.2
marshmallow==3.19.0
marshmallow_sqlalchemy==0.28.1
PyJWT==2.6.0
flask-migrate==4.0.0
flask-sqlalchemy==3.0.2
# mysql-connector-python==8.0.31